chore(controllers): remove error handling on controllers

+++ integrations/install-deps.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
-apk update \
+
+if which jq >/dev/null; then
+    echo
+else
+  apk update \
 	&& apk add jq curl \
 	&& rm -rf /var/cache/apk/*
+fi
